+++ client/src/components/interface/_styling/iconButton.css
@@ -1,6 +1,6 @@
 .icon-button {
-  width: 35px !important;
-  height: 35px !important;
+  width: 35px;
+  height: 35px;
   background-color: var(--red);
   border: 0px solid;
   border-radius: var(--border-radius);
