+++ client/src/components/interface/_styling/segmentedControl.css
@@ -1,5 +1,6 @@
 .segmented-control {
-  margin: 0.2em auto;
+  margin: 0;
   border-radius: var(--border-radius);
   font-size: var(--fz-xs);
+  height: 44px !important;
 }
