+++ client/src/pages/dashboard/_styling/dashboard.css
@@ -8,7 +8,7 @@
 
 #dashboard #content {
   padding: 10px;
-  margin-top: 45px;
+  margin-top: 54px;
   width: 100%;
   font-size: 50px;
   min-height: calc(100vh - 66px);
