+++ client/src/pages/dashboard/historical/data/_styling/dataView.css
@@ -1,0 +1,7 @@
+/*
+Copyright Schulich Racing, FSAE
+Written by Justin Tijunelis
+*/
+
+#data-view {
+}
