--- client/src/pages/dashboard/historical/plots.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Plots: React.FC = () => {
-  return <div id="plots">Plots</div>;
-};
-
-export default Plots;
