+++ client/src/pages/dashboard/historical/plots/plotView.tsx
@@ -1,0 +1,8 @@
+// Copyright Schulich Racing, FSAE
+// Written by Justin Tijunelis
+
+const PlotView: React.FC = () => {
+  return <div></div>;
+};
+
+export default PlotView;
