--- client/src/pages/dashboard/streaming/sensorSelection.tsx
@@ -1,8 +1,0 @@
-// Copyright Schulich Racing FSAE
-// Written by Justin Tijunelis
-
-const SensorSelection: React.FC = () => {
-  return <div></div>;
-};
-
-export default SensorSelection;
